requirements.txt
azure.storage.blob
boto3
google-cloud-speech
google-cloud-storage
matplotlib
omegaconf
protobuf
pvfalcon
pyannote.audio
pyannote.metrics
requests
simple-diarizer
soundfile
torch
tqdm